+++ store/iterator.go
@@ -25,9 +25,9 @@ import (
 //
 type Iterator struct {
 	ctx      context.Context
-	items    chan *KV
+	items    chan KV
 	errorCh  chan error
-	lastItem *KV
+	lastItem KV
 	err      error
 	once     sync.Once
 }
@@ -36,7 +36,7 @@ type Iterator struct {
 func NewIterator(ctx context.Context) *Iterator {
 	return &Iterator{
 		ctx:     ctx,
-		items:   make(chan *KV, 100),
+		items:   make(chan KV, 100),
 		errorCh: make(chan error, 1),
 	}
 }
@@ -65,7 +65,7 @@ func (it *Iterator) Next() bool {
 	return true
 }
 
-func (it *Iterator) Item() *KV {
+func (it *Iterator) Item() KV {
 	return it.lastItem
 }
 
@@ -76,7 +76,7 @@ func (it *Iterator) Err() error {
 //
 // Results gathering primitives
 //
-func (it *Iterator) PushItem(res *KV) bool {
+func (it *Iterator) PushItem(res KV) bool {
 	select {
 	case <-it.ctx.Done():
 		it.PushError(it.ctx.Err())